mesa: Update mesa and mesa-demos to support Sololite.
SoloLite should use mesa software rendering GL since
GPU for this machine does not have hardware accelerated GL.

+++ meta-fsl-arm/recipes-graphics/mesa/mesa_%.bbappend
@@ -2,9 +2,19 @@ PACKAGECONFIG_remove_mx5 = "egl gles"
 PROVIDES_remove_mx5 = "virtual/libgles1 virtual/libgles2 virtual/egl"
 
 PACKAGECONFIG_remove_mx6 = "egl gles"
-PROVIDES_remove_mx6 = "virtual/libgl virtual/libgles1 virtual/libgles2 virtual/egl"
+
+# i.MX6SL uses mesa software rendering
+
+PROVIDES_remove_mx6 = "virtual/libgles1 virtual/libgles2 virtual/egl"
+PROVIDES_remove_mx6q = "virtual/libgl"
+PROVIDES_remove_mx6dl = "virtual/libgl"
+
+USE_VIV_LIBGL = "yes"
+USE_VIV_LIBGL_mx6sl = "no"
 
 # FIXME: Dirty hack to allow use of Vivante GPU libGL binary
 do_install_append_mx6 () {
-    rm -f ${D}${libdir}/libGL.*
+    if [ "${USE_VIV_LIBGL}" = "yes" ]; then
+        rm -f ${D}${libdir}/libGL.*
+    fi
 }
